requirements.txt
streamlit
pdfplumber
pymupdf
google-generativeai
gspread
python-dotenv
pandas
google-auth
oauth2client
